requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langgraph
openai
langchain-openai
pypdf
python-docx
reportlab
beanie
python-jose
passlib[bcrypt]
